init profiles app: refactor models, views, tests, utils of users app
