+++ english_diary/english_diary/settings/partials/base.py
@@ -36,6 +36,7 @@ INSTALLED_APPS = [
 
     'users',
     'diaries',
+    'profiles',
 ]
 
 MIDDLEWARE_CLASSES = [
