+++ english_diary/english_diary/templates/emails/email_verification.html
@@ -68,7 +68,7 @@ table.vb-container.fullwidth {padding-left: 0;padding-right: 0;}</style><style t
         <table width="570" border="0" cellpadding="0" cellspacing="18" class="vb-container fullpad" bgcolor="#ffffff" style="border-collapse: separate;border-spacing: 18px;padding-left: 0;padding-right: 0;width: 100%;max-width: 570px;background-color: #fff;"><tbody><tr><td valign="top" bgcolor="#ffffff" align="center" style="background-color: #ffffff;">
 
               <table cellpadding="0" border="0" align="center" cellspacing="0" class="mobile-full"><tbody><tr><td width="auto" valign="middle" bgcolor="#e53935" align="center" height="50" style="font-size: 15px; font-family: Arial, Helvetica, sans-serif; color: #fff; font-weight: normal; padding-left: 14px; padding-right: 14px; background-color: #e53935; border-radius: 4px;">
-                              <a target="_new" href="https://engdiary.dsds101.info/verify/{{ verification_key }}/" style="text-decoration: none; color: #fff; font-weight: normal;">이메일 인증하기</a>
+                              <a target="_new" href="https://engdiary.dsds101.info/account/verify/{{ verification_key }}/" style="text-decoration: none; color: #fff; font-weight: normal;">이메일 인증하기</a>
                   </td>
                 </tr></tbody></table></td>
           </tr></tbody></table></div>
