+++ english_diary/profiles/models/__init__.py
@@ -1,0 +1,1 @@
+from .profile import Profile
