+++ english_diary/profiles/views/key_expires.py
@@ -2,7 +2,7 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.core.urlresolvers import reverse
 from django.views.generic import View
 
-from users.models import Profile
+from profiles.models import Profile
 
 
 class KeyExpiresView(View):
