+++ english_diary/users/models/__init__.py
@@ -1,2 +1,1 @@
 from .user import User
-from .profile import Profile
