+++ english_diary/users/templates/users/key_expires.html
@@ -8,7 +8,7 @@
             <div class="row">
                 <div class="col-xl-12 col-sm-6 col-sm-offset-3 col-md-4 col-md-offset-4 col-lg-4 col-lg-offset-4">
                     <div class="login-card">
-                        <form action="{% url "users:renew_key" %}" method="POST">
+                        <form action="{% url "profiles:renew_key" %}" method="POST">
                             {% csrf_token %}
                             <input type="hidden" name="verification_key" value="{{ verification_key }}">
                                 <p>인증 메일의 유효기간이 지났습니다.</p>
