+++ english_diary/users/tests/test_models.py
@@ -1,9 +1,30 @@
 from django.contrib.auth import get_user_model
 
-from .base import UserBaseTestCase
+from core.tests.base import BaseTestCase
 
 
-class UserModelTestCase(UserBaseTestCase):
+class UserModelTestCase(BaseTestCase):
+
+    def setUp(self):
+        super(UserModelTestCase, self).setUp()
+
+        # Create sample diaries
+        self.user.diary_set.create(
+            datetime="2016/07/02",
+            content="Today it was very hot!",
+        )
+        self.user.diary_set.create(
+            datetime="2016/07/15",
+            content="I went to the concert. And I had dinner with my friends",
+        )
+        self.user.diary_set.create(
+            datetime="2016/08/11",
+            content="I watched a action movie with my family. It was so exciting.",
+        )
+        self.user.diary_set.create(
+            datetime="2016/08/29",
+            content="I was so tired today. I don`t want to do no more",
+        )
 
     def test_check_username_in_user_model(self):
