+++ english_diary/users/views/__init__.py
@@ -2,6 +2,3 @@ from .signin import SigninView
 from .signout import SignoutView
 from .signup import SignupView
 from .mypage import MypageView
-from .email_verification import EmailVerificationView
-from .key_expires import KeyExpiresView
-from .renew_key import RenewKeyView
